Rename testutils project and adapt namespace
