+++ bundles/extensions/constructionsimulation/tools.vitruv.extensions.constructionsimulation.util/META-INF/MANIFEST.MF
@@ -10,7 +10,7 @@ Require-Bundle: org.eclipse.ui,
  tools.vitruv.framework.vsum,
  tools.vitruv.framework.change,
  tools.vitruv.framework.change.processing,
- tools.vitruv.framework.tests.util
+ tools.vitruv.testutils
 Import-Package: org.apache.commons.io,
  org.eclipse.emf.common.util,
  org.eclipse.emf.ecore,
