+++ bundles/extensions/constructionsimulation/tools.vitruv.extensions.constructionsimulation/META-INF/MANIFEST.MF
@@ -16,7 +16,7 @@ Require-Bundle: org.junit,
  org.eclipse.ui.workbench,
  org.eclipse.emf.ecore,
  org.eclipse.emf.emfstore.common,
- tools.vitruv.framework.tests.util,
+ tools.vitruv.testutils,
  tools.vitruv.framework.vsum,
  tools.vitruv.framework.util,
  tools.vitruv.framework.userinteraction,
