+++ bundles/testutils/tools.vitruv.testutils/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>tools.vitruv.framework.tests.util</name>
+	<name>tools.vitruv.testutils</name>
 	<comment></comment>
 	<projects>
 	</projects>
