+++ bundles/testutils/tools.vitruv.testutils/src/tools/vitruv/testutils/matchers/ModelMatchers.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.tests.matchers
+package tools.vitruv.testutils.matchers
 
 import org.eclipse.emf.ecore.EObject
 import org.hamcrest.Description
