+++ tests/dsls/tools.vitruv.dsls.commonalities.tests/META-INF/MANIFEST.MF
@@ -11,7 +11,7 @@ Require-Bundle: org.eclipse.xtext.testing,
  tools.vitruv.dsls.commonalities,
  org.hamcrest.core;bundle-version="1.3.0",
  org.eclipse.jdt.core;bundle-version="3.13.0",
- tools.vitruv.framework.tests.util;bundle-version="0.2.0",
+ tools.vitruv.testutils;bundle-version="0.2.0",
  tools.vitruv.framework.testutils.domains;bundle-version="0.2.0",
  tools.vitruv.extensions.dslsruntime.reactions;bundle-version="0.2.0",
  org.eclipse.core.runtime;bundle-version="3.13.0"
