+++ tests/dsls/tools.vitruv.dsls.commonalities.ui.tests/META-INF/MANIFEST.MF
@@ -17,7 +17,7 @@ Require-Bundle: tools.vitruv.dsls.commonalities.ui,
  tools.vitruv.dsls.commonalities,
  org.eclipse.emf;bundle-version="2.6.0",
  org.eclipse.jdt.core;bundle-version="3.13.0",
- tools.vitruv.framework.tests.util;bundle-version="0.2.0",
+ tools.vitruv.testutils;bundle-version="0.2.0",
  tools.vitruv.framework.testutils.domains;bundle-version="0.2.0",
  org.eclipse.xtext.ui;bundle-version="2.12.0",
  org.eclipse.pde.core;bundle-version="3.11.100",
