+++ tests/dsls/tools.vitruv.dsls.commonalities.ui.tests/src/tools/vitruv/dsls/commonalities/ui/executiontests/CommonalitiesExecutionTest.xtend
@@ -1,6 +1,6 @@
 package tools.vitruv.dsls.commonalities.ui.executiontests
 
-import tools.vitruv.framework.tests.VitruviusApplicationTest
+import tools.vitruv.testutils.VitruviusApplicationTest
 
 import static extension edu.kit.ipd.sdq.commons.util.java.lang.IterableUtil.*
