+++ tests/dsls/tools.vitruv.dsls.mappings.tests.addressesXrecipients/META-INF/MANIFEST.MF
@@ -5,23 +5,23 @@ Bundle-SymbolicName: tools.vitruv.dsls.mappings.tests.addressesXrecipients
 Bundle-Version: 0.2.0.qualifier
 Bundle-Vendor: vitruv.tools
 Require-Bundle: org.apache.log4j,
-  com.google.guava,
-  org.eclipse.emf.ecore,
-  org.eclipse.core.resources,
-  org.eclipse.xtend.lib,
-  org.eclipse.xtend.lib.macro,
-  tools.vitruv.framework.util,
-  tools.vitruv.framework.change,
-  tools.vitruv.framework.change.processing,
-  tools.vitruv.framework.correspondence,
-  tools.vitruv.framework.userinteraction,
-  tools.vitruv.extensions.dslsruntime.reactions,
-  edu.kit.ipd.sdq.metamodels.addresses,
-  edu.kit.ipd.sdq.metamodels.recipients,
-  tools.vitruv.demo.domains.addresses,
-  tools.vitruv.demo.domains.recipients,
-  tools.vitruv.extensions.dslsruntime.mappings,
-  tools.vitruv.framework.tests.util,
-  org.junit
+ com.google.guava,
+ org.eclipse.emf.ecore,
+ org.eclipse.core.resources,
+ org.eclipse.xtend.lib,
+ org.eclipse.xtend.lib.macro,
+ tools.vitruv.framework.util,
+ tools.vitruv.framework.change,
+ tools.vitruv.framework.change.processing,
+ tools.vitruv.framework.correspondence,
+ tools.vitruv.framework.userinteraction,
+ tools.vitruv.extensions.dslsruntime.reactions,
+ edu.kit.ipd.sdq.metamodels.addresses,
+ edu.kit.ipd.sdq.metamodels.recipients,
+ tools.vitruv.demo.domains.addresses,
+ tools.vitruv.demo.domains.recipients,
+ tools.vitruv.extensions.dslsruntime.mappings,
+ tools.vitruv.testutils,
+ org.junit
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: mir.reactions
