+++ tests/dsls/tools.vitruv.dsls.reactions.tests/META-INF/MANIFEST.MF
@@ -6,35 +6,35 @@ Bundle-Version: 0.2.0.qualifier
 Bundle-Vendor: vitruv.tools
 Export-Package: tools.vitruv.dsls.reactions.tests;x-internal=true
 Require-Bundle: org.junit,
-  org.apache.log4j,
-  com.google.guava,
-  org.eclipse.core.resources,
-  org.eclipse.emf.ecore,
-  org.eclipse.xtext.xbase.junit,
-  org.eclipse.xtext.xbase.lib,
-  org.eclipse.xtext.builder,
-  org.eclipse.xtend.lib,
-  org.eclipse.xtend.lib.macro,
-  tools.vitruv.framework.util,
-  tools.vitruv.framework.tests.util,
-  tools.vitruv.framework.tests.util.metamodels,
-  tools.vitruv.framework.change,
-  tools.vitruv.framework.change.processing,
-  tools.vitruv.framework.correspondence,
-  tools.vitruv.framework.userinteraction,
-  tools.vitruv.framework.metamodel,
-  tools.vitruv.framework.tuid,
-  tools.vitruv.dsls.reactions,
-  tools.vitruv.dsls.reactions.ui,
-  tools.vitruv.extensions.dslsruntime.reactions,
-  org.objectweb.asm;bundle-version="[5.0.1,6.0.0)";resolution:=optional,
-  tools.vitruv.framework.testutils.domains,
-  org.eclipse.xtext.testing;bundle-version="2.12.0",
-  org.eclipse.xtext.xbase.testing,
-  tools.vitruv.extensions.emf;bundle-version="0.2.0",
-  org.eclipse.xtext.xtext.generator;bundle-version="2.12.0",
-  org.eclipse.jdt.core;bundle-version="3.13.0",
-  tools.vitruv.extensions.dslsruntime.mappings
+ org.apache.log4j,
+ com.google.guava,
+ org.eclipse.core.resources,
+ org.eclipse.emf.ecore,
+ org.eclipse.xtext.xbase.junit,
+ org.eclipse.xtext.xbase.lib,
+ org.eclipse.xtext.builder,
+ org.eclipse.xtend.lib,
+ org.eclipse.xtend.lib.macro,
+ tools.vitruv.framework.util,
+ tools.vitruv.testutils,
+ tools.vitruv.framework.tests.util.metamodels,
+ tools.vitruv.framework.change,
+ tools.vitruv.framework.change.processing,
+ tools.vitruv.framework.correspondence,
+ tools.vitruv.framework.userinteraction,
+ tools.vitruv.framework.metamodel,
+ tools.vitruv.framework.tuid,
+ tools.vitruv.dsls.reactions,
+ tools.vitruv.dsls.reactions.ui,
+ tools.vitruv.extensions.dslsruntime.reactions,
+ org.objectweb.asm;bundle-version="[5.0.1,6.0.0)";resolution:=optional,
+ tools.vitruv.framework.testutils.domains,
+ org.eclipse.xtext.testing;bundle-version="2.12.0",
+ org.eclipse.xtext.xbase.testing,
+ tools.vitruv.extensions.emf;bundle-version="0.2.0",
+ org.eclipse.xtext.xtext.generator;bundle-version="2.12.0",
+ org.eclipse.jdt.core;bundle-version="3.13.0",
+ tools.vitruv.extensions.dslsruntime.mappings
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.hamcrest.core,
   org.junit;version="4.5.0",
