+++ tests/extensions/emfdomain/tools.vitruv.extensions.emf.monitorededitor.tests/META-INF/MANIFEST.MF
@@ -26,7 +26,7 @@ Require-Bundle: org.eclipse.ui,
  tools.vitruv.framework.change.processing,
  tools.vitruv.framework.tuid,
  tools.vitruv.framework.vsum,
- tools.vitruv.framework.tests.util
+ tools.vitruv.testutils
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Export-Package: tools.vitruv.domains.emf.monitorededitor.test.emfmodels.advancedfeatures,
