+++ tests/extensions/emfdomain/tools.vitruv.extensions.emf.monitorededitor.tests/src/tools/vitruv/domains/emf/monitorededitor/test/utils/BasicTestCase.java
@@ -16,7 +16,7 @@ import java.net.URL;
 import org.apache.log4j.Logger;
 import org.eclipse.emf.common.util.URI;
 
-import tools.vitruv.framework.tests.util.TestUtil;
+import tools.vitruv.testutils.util.TestUtil;
 
 public class BasicTestCase {
     public static final String PLUGIN_NAME = "tools.vitruv.extensions.emf.monitorededitor.tests";
