+++ tests/framework/tools.vitruv.framework.vsum.tests/META-INF/MANIFEST.MF
@@ -11,7 +11,7 @@ Require-Bundle: org.junit,
  tools.vitruv.framework.util,
  org.eclipse.core.runtime,
  org.eclipse.core.resources,
- tools.vitruv.framework.tests.util,
+ tools.vitruv.testutils,
  tools.vitruv.framework.tests.util.metamodels,
  org.apache.commons.lang,
  com.google.guava,
