+++ tests/framework/tools.vitruv.framework.vsum.tests/src/tools/vitruv/framework/tests/vsum/VsumTest.java
@@ -18,14 +18,14 @@ import pcm_mockup.Repository;
 import tools.vitruv.framework.domains.AbstractVitruvDomain;
 import tools.vitruv.framework.domains.VitruvDomain;
 import tools.vitruv.framework.domains.VitruviusProjectBuilderApplicator;
-import tools.vitruv.framework.tests.TestUserInteractor;
-import tools.vitruv.framework.tests.VitruviusTest;
-import tools.vitruv.framework.tests.util.TestUtil;
 import tools.vitruv.framework.tuid.AttributeTuidCalculatorAndResolver;
 import tools.vitruv.framework.util.bridges.EMFBridge;
 import tools.vitruv.framework.util.datatypes.ModelInstance;
 import tools.vitruv.framework.util.datatypes.VURI;
 import tools.vitruv.framework.vsum.InternalVirtualModel;
+import tools.vitruv.testutils.TestUserInteractor;
+import tools.vitruv.testutils.VitruviusTest;
+import tools.vitruv.testutils.util.TestUtil;
 import uml_mockup.UClass;
 import uml_mockup.UPackage;
 import uml_mockup.Uml_mockupFactory;
